adicionado alterações da correção

+++ questao_2.py
@@ -1,5 +1,5 @@
-a=int(input("Digite um valor:"))
-b=int(input("digite outro valor:"))
+a=float(input("Digite um valor:"))
+b=float(input("digite outro valor:"))
 c=a+b
 d=a*b
 print(f"O a soma dos valores informado é:{c}")
